Move the StimulusReflex methods that are assigned to controller
By moving the StimulusReflex methods assigned to the controller it
decouples the way a StimulusReflex controller is registered to the
methods.

This lets a developer outside of the Rails ecosystem re-use those
methods and make their own "register" method without needing to
install actioncable.
